+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.1 - Unreleased
 ### Added:
 - bundle install when running reissue
+### Fixed:
+- bug in tests loading the changelog fixture
 
 ## 0.1.0 - 2024-04-11
 ### Added: