+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.1 - Unreleased
 ### Added:
 - bundle install when running reissue
+- Reissue::Parser class to parse the changelog file
 ### Fixed:
 - bug in tests loading the changelog fixture
+### Removed:
+- dependency on the keepachangelog gem
 
 ## 0.1.0 - 2024-04-11
 ### Added: